working bundle script

+++ libtbx/auto_build/make_bundle.py
@@ -13,10 +13,15 @@ CCTBX-<version>/cctbx_project/
 
 plus any number of module directories in the top level.  The resulting bundle
 will be named bundle-<version>-<mtype>.tar.gz.
+
+Since the base modules take an especially long time to compile, they are now
+part of a separate bundle.  This will allow re-use of precompiled base packages
+with the latest source, which should speed up installer generation.
 """
 
 from __future__ import division
 from optparse import OptionParser
+from cStringIO import StringIO
 import os.path as op
 import shutil
 import time
@@ -24,8 +29,9 @@ import os
 import sys
 # local imports
 # XXX HACK
-libtbx_path = op.abspath(op.dirname(op.dirname(__file__)))
+libtbx_path = op.abspath(op.dirname(op.dirname(op.dirname(__file__))))
 if (not libtbx_path in sys.path) :
+  print libtbx_path
   sys.path.append(libtbx_path)
 from libtbx.auto_build.installer_utils import *
 from libtbx.auto_build import rpath
@@ -44,14 +50,25 @@ def run (args, out=sys.stdout) :
     help="Subdirectories to ignore", default="")
   parser.add_option("--remove_src", dest="remove_src", action="store",
     help="Remove compiled source files (.h, .cpp, etc.)", default=False)
+  parser.add_option("--keep_base", dest="keep_base", action="store",
+    help="Keep base packages with main bundle", default=False)
+  parser.add_option("--dest", dest="dest", action="store",
+    help="Destination directory for bundle tarfiles", default=None)
   options, args = parser.parse_args(args)
   target_dir = args[0]
   assert op.isdir(target_dir), target_dir
   os.chdir(options.tmp_dir)
+  pkg_dir = op.basename(target_dir)
   build_dir = op.join(target_dir, "build", options.mtype)
-  tmp_dir = op.join(options.tmp_dir, "tmp_%s" % options.version)
+  print >> out, "Setting rpath in shared libraries..."
+  stdout_old = sys.stdout
+  sys.stdout = StringIO()
+  rpath.run([build_dir])
+  sys.stdout = stdout_old
+  # create temp dir
+  tmp_dir = op.join(options.tmp_dir, "%s_tmp" % pkg_dir)
   assert op.isdir(build_dir), build_dir
-  if os.exists(tmp_dir) :
+  if op.exists(tmp_dir) :
     shutil.rmtree(tmp_dir)
   os.mkdir(tmp_dir)
   os.chdir(tmp_dir)
@@ -72,15 +89,6 @@ def run (args, out=sys.stdout) :
     full_path = op.join(build_dir, dir_name)
     assert op.isdir(full_path)
     copy_tree(full_path, op.join(tmp_build_dir, dir_name))
-  # copy over build executable directories
-  for file_name in os.listdir(build_dir) :
-    full_path = op.join(build_dir)
-    if op.isdir(full_path) :
-      module_name = file_name
-      for file_name in os.listdir(full_path) :
-        if (file_name == "exe") :
-          copy_tree(op.join(full_path, file_name),
-                    op.join(tmp_build_dir, module_name, file_name))
   # remove unnecessary base directories/files
   for dir_name in [
       "base/bin/gtk-demo",
@@ -91,11 +99,56 @@ def run (args, out=sys.stdout) :
       "base/share/locale",
     ] :
     full_path = op.join(tmp_build_dir, dir_name)
-    shutil.rmtree(full_path)
+    if op.exists(full_path) :
+      shutil.rmtree(full_path)
   # XXX what about base/include?
+  # copy over build executable directories
+  for file_name in os.listdir(build_dir) :
+    full_path = op.join(build_dir)
+    if op.isdir(full_path) :
+      module_name = file_name
+      for file_name in os.listdir(full_path) :
+        if (file_name == "exe") :
+          copy_tree(op.join(full_path, file_name),
+                    op.join(tmp_build_dir, module_name, file_name))
   # delete unnecessary files
   find_and_delete_files(tmp_dir, file_ext=".pyc")
+  find_and_delete_files(tmp_dir, file_ext=".o")
   find_and_delete_files(tmp_dir, file_ext=".pyo")
   find_and_delete_files(tmp_dir, file_name=".sconsign")
   find_and_delete_files(tmp_dir, file_name="CVS")
   find_and_delete_files(tmp_dir, file_name=".svn")
+  if (options.remove_src) :
+    find_and_delete_files(tmp_dir, file_ext=".cpp")
+    find_and_delete_files(tmp_dir, file_ext=".hpp")
+    find_and_delete_files(tmp_dir, file_ext=".cc")
+    find_and_delete_files(tmp_dir, file_ext=".c")
+    find_and_delete_files(tmp_dir, file_ext=".h")
+  # TODO strip objects?
+  os.chdir(tmp_dir)
+  # create base bundle
+  if (not options.keep_base) :
+    base_dir = op.join("build", options.mtype, "base")
+    base_tarfile = "../base-%(version)s-%(mtype)s.tar.gz" % \
+      {"version":options.version, "mtype":options.mtype}
+    call("tar -czf %(tarfile)s %(base)s" %
+      {"tarfile":base_tarfile, "base":base_dir}, log=out)
+    shutil.rmtree(base_dir)
+    assert op.isfile(base_tarfile)
+    if (options.dest is not None) :
+      shutil.move(base_tarfile, options.dest)
+      base_tarfile = op.join(options.dest, op.basename(base_tarfile))
+    print >> out, "  created base bundle %s" % base_tarfile
+  # create the product bundle
+  pkg_tarfile = "../bundle-%(version)s-%(mtype)s.tar.gz" % \
+    {"version":options.version, "mtype":options.mtype}
+  call("tar -czf %(tarfile)s ." % {"tarfile":pkg_tarfile}, log=out)
+  assert op.isfile(pkg_tarfile)
+  if (options.dest is not None) :
+    shutil.move(pkg_tarfile, options.dest)
+    pkg_tarfile = op.join(options.dest, op.basename(pkg_tarfile))
+  print >> out, "  created bundle %s" % pkg_tarfile
+  shutil.rmtree(tmp_dir)
+
+if (__name__ == "__main__") :
+  run(sys.argv[1:])
